requirements.txt
# Python dependencies for YOLOv8 fire and smoke detection
ultralytics>=8.0.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
